+++ apps/mobile/src/lib/db/migrations/migrations.js
@@ -1,16 +1,8 @@
 // This file is required for Expo/React Native SQLite migrations - https://orm.drizzle.team/quick-sqlite/expo
 
-import journal from './meta/_journal.json';
-import m0000 from './0000_abandoned_quasimodo.sql';
-import m0001 from './0001_flashy_xorn.sql';
-import m0002 from './0002_next_sharon_ventura.sql';
+import journal from "./meta/_journal.json";
 
-  export default {
-    journal,
-    migrations: {
-      m0000,
-m0001,
-m0002
-    }
-  }
-  
+export default {
+  journal,
+  migrations: {},
+};